Generify ajax pagination JS and add first, last links
Instead of using the ajax pagination file purely for the user
directory, we want to use it for any page on the site. In this commit
we therefore generify the AjaxPagination JS file so that it can be
added for any set of data. We also have added the functionality for
the 'first' and 'last' button links (middle page links still to come).