Generify the ajax pagination functionality
This commit aims to further generify the ajax pagination functionality
so that it can be used on pages other than just the user directory.
This includes changes to JS and PHP files in order to make the
pagination as easy to integrate into other pages as possible. Note
that we have also refactored the 'grid' and 'list' functionality a bit in
order to use that functionality on other pages as well.