Fix (enable) user directory list view button on search
When 'search' is clicked on the user directory page, we want both of
the view buttons to be enabled. The code was only enabling the grid
(default) view button instead of both buttons. This button fixes the
issue by enabling the view button correctly after the search button
is clicked.