Rename 'item' table to 'item_type'
Our item design is taking a different direction. We want a
table with the item types and a separate table for the
items in order to speed up the checkin/checkout
updates. We will update the collection child tables in the
database in the near future so as to point to actual items
and not just the item types.